Refactor: move files to packages
