+++ app/src/main/java/earth/satunes/ui/components/dialog/MediaSelectionDialog.kt
@@ -27,7 +27,6 @@ package earth.satunes.ui.components.dialog
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -35,7 +34,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import earth.satunes.R
 import earth.satunes.database.models.Media
-import earth.satunes.ui.components.cards.dialog.DialogOption
 import earth.satunes.ui.components.forms.MediaSelectionForm
 import earth.satunes.ui.components.texts.NormalText
