+++ app/src/main/java/earth/satunes/ui/components/dialog/PlaylistOptionsDialog.kt
@@ -36,7 +36,6 @@ import androidx.compose.ui.unit.dp
 import earth.satunes.R
 import earth.satunes.database.models.relations.PlaylistWithMusics
 import earth.satunes.database.models.tables.Playlist
-import earth.satunes.ui.components.cards.dialog.DialogOption
 import earth.satunes.ui.components.texts.NormalText
 import earth.satunes.icons.SatunesIcons
