+++ app/src/main/java/earth/satunes/ui/views/settings/UpdatesView.kt
@@ -38,13 +38,13 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import earth.satunes.R
-import earth.satunes.internet.UpdateAvailableStatus
-import earth.satunes.internet.UpdateAvailableStatus.AVAILABLE
-import earth.satunes.internet.UpdateAvailableStatus.CANNOT_CHECK
-import earth.satunes.internet.UpdateAvailableStatus.UNDEFINED
-import earth.satunes.internet.UpdateAvailableStatus.UP_TO_DATE
-import earth.satunes.internet.UpdateCheckManager
-import earth.satunes.internet.UpdateCheckManager.getCurrentVersion
+import earth.satunes.internet.updates.UpdateAvailableStatus
+import earth.satunes.internet.updates.UpdateAvailableStatus.AVAILABLE
+import earth.satunes.internet.updates.UpdateAvailableStatus.CANNOT_CHECK
+import earth.satunes.internet.updates.UpdateAvailableStatus.UNDEFINED
+import earth.satunes.internet.updates.UpdateAvailableStatus.UP_TO_DATE
+import earth.satunes.internet.updates.UpdateCheckManager
+import earth.satunes.internet.updates.UpdateCheckManager.getCurrentVersion
 import earth.satunes.ui.components.LoadingCircle
 import earth.satunes.ui.components.buttons.updates.CheckUpdateButton
 import earth.satunes.ui.components.settings.UpdateAvailable
