+++ internet/src/main/java/earth/satunes/internet/updates/UpdateAvailableStatus.kt
@@ -23,23 +23,14 @@
  *  PS: I don't answer quickly.
  */
 
-package earth.satunes.internet
-
-import android.app.Activity
-import android.app.DownloadManager
-import android.content.BroadcastReceiver
-import android.content.Context
-import android.content.Intent
-import earth.satunes.internet.utils.showToastOnUiThread
+package earth.satunes.internet.updates
 
 /**
- * @author Antoine Pirlot on 14/04/2024
+ * @author Antoine Pirlot on 11/04/2024
  */
-internal object DownloadReceiver : BroadcastReceiver() {
-    override fun onReceive(context: Context, intent: Intent) {
-        if (intent.action == DownloadManager.ACTION_DOWNLOAD_COMPLETE) {
-            UpdateCheckManager.downloadStatus.value = APKDownloadStatus.DOWNLOADED
-            showToastOnUiThread(context = context, activity = Activity(), context.getString(R.string.downloaded))
-        }
-    }
+enum class UpdateAvailableStatus(var updateLink: String? = null) {
+    AVAILABLE,
+    CANNOT_CHECK,
+    UNDEFINED,
+    UP_TO_DATE,
 }
